# tests: increase timeout for nrf51dk