--- a/tests/pkg_libcose/tests/01-run.py
+++ b/tests/pkg_libcose/tests/01-run.py
@@ -14,8 +14,8 @@ from testrunner import TIMEOUT as DEFAULT_TIMEOUT
 
 BOARD = os.environ['BOARD']
 # on real hardware, this test application can take several minutes to
-# complete (~4min on microbit)
-TIMEOUT = 300 if BOARD != 'native' else DEFAULT_TIMEOUT
+# complete (>5min on nrf51dk)
+TIMEOUT = 400 if BOARD != 'native' else DEFAULT_TIMEOUT
 
 
 if __name__ == "__main__":