--- a/tests/pkg_wolfssl/tests/01-run.py
+++ b/tests/pkg_wolfssl/tests/01-run.py
@@ -11,9 +11,11 @@ BOARD = os.environ.get("BOARD", "native")
 # Increase timeout on "real" hardware
 # ED25519 takes +160s on samr21-xpro
 # ED25519 takes +230s on nucleo-l073rz
-TEST_TIMEOUT = 300 if BOARD != 'native' else DEFAULT_TIMEOUT
+# ED25519 takes +500s on nrf51dk
+TEST_TIMEOUT = 600 if BOARD != 'native' else DEFAULT_TIMEOUT
 # ECDSA 256 takes +30s on samr21-xpro
-BENCH_TIMEOUT = 30 if BOARD != 'native' else DEFAULT_TIMEOUT
+# ECDSA 256 takes +40s on nrf51dk
+BENCH_TIMEOUT = 40 if BOARD != 'native' else DEFAULT_TIMEOUT
 
 
 def _wait_for_test(child):