--- a/tests/suit_manifest/tests/01-run.py
+++ b/tests/suit_manifest/tests/01-run.py
@@ -15,7 +15,8 @@ def testfunc(child):
     board = os.environ['BOARD']
     # Increase timeout on "real" hardware
     # 16 seconds on `samr21-xpro`
-    timeout = 30 if board != 'native' else -1
+    # >50 seconds on `nrf51dk`
+    timeout = 60 if board != 'native' else -1
     child.expect(r"OK \(\d+ tests\)", timeout=timeout)